ASN annotator updated to the refactored loading method
